feat: 全面完善ERP进销存系统功能、UI美化、Bug修复、Docker/FPK部署
## 🆕 新增功能
- 销售单取消功能 (sale cancel action)
- 采购单一键确认入库 (purchase confirm_inbound)
- StockInItem入库明细模型 + 序列化器 + 迁移
- SaleItem自动计算金额 (save方法)
- 收付款单号正确生成 (generate_payment_no)
- Token刷新修复 (request._full_data替代不可变request.data)
- LogViewSet权限修复 (添加ModulePermission)
- 登出操作日志记录
- 操作日志Action类型扩展(取消/确认/确认入库/更新状态)

## 🐛 Bug修复
- PrintTemplateViewSet缺少permission_classes(500错误)
- GoodsViewSet.search_fields移除不存在的brand字段
- CategoryViewSet删除重复的batch_update_sort方法
- SaleOrderViewSet添加删除保护(状态检查)
- useInventoryLog.js添加缺失的API导入
- useSaleOrders.js添加form.id(编辑功能修复)
- usePayments.js添加handleDialogClose
- useGoods.js添加priceErrors/handlePriceInput/handlePriceBlur导出
- SaleOrders.vue添加handleStockOutFromView
- 4个composable添加onUnmounted事件监听器清理
- usePurchaseOrders.js删除重复availableGoods和死代码selectedGoods
- 销售出库允许partial状态(与采购入库一致)
- 销售出库添加库存预检查(友好中文提示)
- 出库单号生成改为generate_stock_out_no(修复UNIQUE冲突)
- Login.vue表单验证与API调用分离
- SaleOrderForm修复totalAmount.toFixed错误

## 🎨 UI美化
- 全局CSS变量精简 + Element Plus主题增强
- page-common.css统一设计规范(渐变工具栏/统计卡片动画/状态徽章)
- 7个页面从独立样式转换为common-page统一模式
- Layout添加页面切换过渡动画
- Dashboard统计卡片悬浮动效
- Login玻璃拟态卡片 + 多色渐变背景
- SaleOrderView添加完整scoped样式

## 🚀 部署
- NAS Docker Compose部署(支持SQLite/MySQL)
- FPK完整打包(fnos manifest + lifecycle脚本 + 卸载清理)
- entrypoint.sh自动迁移+创建管理员
- .dockerignore + .env环境配置

+++ erp/purchase/models.py
@@ -39,8 +39,8 @@ class PurchaseItem(models.Model):
     """采购明细"""
     order = models.ForeignKey(PurchaseOrder, on_delete=models.CASCADE, related_name='items', verbose_name='采购单')
     goods = models.ForeignKey(Goods, on_delete=models.PROTECT, verbose_name='商品')
-    quantity = models.DecimalField(max_digits=12, decimal_places=2, verbose_name='数量')
-    received_quantity = models.DecimalField(max_digits=12, decimal_places=2, default=0, verbose_name='已入库数量')
+    quantity = models.IntegerField(verbose_name='数量')
+    received_quantity = models.IntegerField(default=0, verbose_name='已入库数量')
     price = models.DecimalField(max_digits=12, decimal_places=2, verbose_name='单价')
     amount = models.DecimalField(max_digits=14, decimal_places=2, verbose_name='金额')
     remark = models.CharField(max_length=200, blank=True, verbose_name='备注')
